Finished MVP Game

+++ public/game.js
@@ -33,6 +33,17 @@ var orangeDinoInterval;
 var purpleDinos;
 var purpleDinoInterval;
 
+var fishes;
+var fishInterval;
+
+var music;
+var musicArray = ['fox', 'gorillaz', 'lucky'];
+var randomSong = Math.floor(Math.random()*3);
+var musicReset = false;
+var musicLoop = true;
+var jumpEffect;
+var deadEffect;
+
 var updatePosition = function(positionArray) {
 
   for(var i=0; i<positionArray.length; i++) {
@@ -40,7 +51,6 @@ var updatePosition = function(positionArray) {
       POS_X = positionArray[i].data.velocity.x;
       POS_Y = positionArray[i].data.velocity.y;
       BUTTON_A = positionArray[i].data.a;
-      console.log(POS_X, POS_Y, BUTTON_A);
       if(RESET && BUTTON_A) {
         GAMECONTEXT.move();
       }
@@ -64,11 +74,31 @@ var state = {
     this.load.image("background", "assets/background.jpg");
     this.load.image("floating", "assets/floating.png");
     this.load.spritesheet("player", "assets/hero.png", 33.16, 49);
+    this.load.spritesheet("fish", "assets/fish.png", 30, 40);
     this.load.spritesheet("orangeDino", "assets/orange-dino.png", 34.5, 42);
     this.load.spritesheet("purpleDino", "assets/purple-dino.png", 118, 150);
     this.load.image('water', 'assets/water.png');
+    this.load.audio('fox', ['assets/sounds/fox.mp3']);
+    this.load.audio('gorillaz', ['assets/sounds/gorillaz.mp3']);
+    this.load.audio('lucky', ['assets/sounds/lucky.mp3']);
+    this.load.audio('dead', ['assets/sounds/dead.mp3']);
+    this.load.audio('jump', ['assets/sounds/jump.mp3']);
   },
   create: function() {
+
+    music = game.add.audio(musicArray[randomSong]);
+    music.play();
+    //   musicReset = false;
+
+    // music.restart();
+    music.onStop.add(function(){
+      if(musicLoop === true) {
+        music.play()
+      }  
+    }, this);
+    // music.onStop.add(music.play);
+    // music.loop = true;
+
     this.physics.startSystem(Phaser.Physics.ARCADE);
 
     this.background = this.add.tileSprite(0,0, this.world.width, this.world.height, 'background');
@@ -94,6 +124,8 @@ var state = {
     purpleDinos = game.add.group();
     orangeDinos = game.add.group();
 
+    fishes = game.add.group();
+
 
     this.scoreText = this.add.text(
       this.world.centerX,
@@ -115,16 +147,17 @@ var state = {
     this.physics.arcade.collide(players, platforms);
     this.physics.arcade.collide(players, orangeDinos, this.setGameOver, null, this);
     this.physics.arcade.collide(players, purpleDinos, this.setGameOver, null, this);
+    this.physics.arcade.collide(players, fishes, this.setGameOver, null, this);
 
 
     if(this.level === 'ground') {
       orangeDinos.forEach(function(o) {
-        if(o.body.x < -100) {
+        if(o && o.body.x < -100) {
           o.kill();
         }
       });
       purpleDinos.forEach(function(p) {
-        if(p.body.x < -150) {
+        if(p && p.body.x < -150) {
           p.kill();
         }
       });
@@ -139,6 +172,11 @@ var state = {
       platforms.forEach(function(p){
         if(p.body.x < -800 || p.body.y < -48 ) {
           p.kill();
+        }
+      });
+      fishes.forEach(function(f) {
+        if(f.body.x < - 100) {
+          f.kill();
         }
       });
     }
@@ -157,6 +195,8 @@ var state = {
 
     if(BUTTON_A && this.player.body.touching.down && !this.player.dead) {
       this.player.body.velocity.y = -600;
+      jumpEffect = game.add.audio('jump');
+      jumpEffect.play();
     }
     
  
@@ -198,6 +238,7 @@ var state = {
     clearInterval(platformFloatingInterval);
     clearInterval(purpleDinoInterval);
     clearInterval(orangeDinoInterval);
+    clearInterval(fishInterval);
 
     clearTimeout(waterTimeout);
     clearTimeout(groundTimeout);
@@ -210,8 +251,22 @@ var state = {
     water.removeAll();
     orangeDinos.removeAll();
     purpleDinos.removeAll();
+    fishes.removeAll();
     this.gameStarted = false;
     this.gameOver = false;
+
+    if(musicReset === true) {
+      musicLoop = true;
+      musicReset = false;
+      music.play();
+
+      music.onStop.add(function(){
+        if(musicLoop === true) {
+          music.play()
+        }  
+      }, this);
+    }
+
 
     // this.input.onDown.removeAll();    
     // this.input.onDown.add(this.move, this);
@@ -297,6 +352,20 @@ var state = {
     this.player.body.velocity.x = 0;
     this.player.body.velocity.y = 0;
     this.player.animations.play('still'); 
+
+    deadEffect = game.add.audio('dead');
+    deadEffect.play();
+
+    if(randomSong === 2) {
+      randomSong = 0;
+    } else {
+      randomSong++
+    }
+    musicReset = true;
+    musicLoop = false;
+    music.stop();
+    music = game.add.audio(musicArray[randomSong]);
+
     // this.input.onDown.removeAll();
     // this.input.onDown.add(this.reset, this);
   },
@@ -372,6 +441,16 @@ var state = {
     this.purpleDino.body.immovable = true;
     this.purpleDino.body.velocity.x = -SPEED - 80;
   },
+  spawnFish: function() {
+    this.fish = fishes.create(700, 650, 'fish');
+    this.fish.animations.add('walk', [0,1], 10, true);
+    this.fish.animations.play('walk');
+    this.physics.arcade.enableBody(this.fish);
+    this.fish.body.immovable = true;
+    this.fish.body.velocity.y = -900;
+    this.fish.body.velocity.x = -SPEED;
+    this.fish.body.gravity.y = GRAVITY;
+  },
   levelWater: function() {
     this.level = 'water';
 
@@ -403,6 +482,9 @@ var state = {
       platformNegativeInterval = setInterval(function(){
         context.spawnNegativePlatform();
       }, 8000);
+      fishInterval = setInterval(function() {
+        context.spawnFish();
+      }, 3500);
     } 
   },
   levelGround: function() {
@@ -413,6 +495,7 @@ var state = {
     clearInterval(platformFallingInterval);
     clearInterval(platformNegativeInterval);
     clearInterval(waterInterval);
+    clearInterval(fishInterval);
 
 
     this.ground = platforms.create(800, game.world.height-64, 'ground');
